separated zod from dto
