+++ package.json
@@ -5,7 +5,6 @@
   "type": "module",
   "scripts": {
     "dev": "tsx watch src/server.ts",
-    "init": "tsx src/migrate.ts",
     "migrate": "node dist/migrate-to-latest.js",
     "build": "rm -rf dist && tsc",
     "build:win": "del dist && del .tsbuildinfo && tsc",
@@ -16,7 +15,8 @@
     "docker:clean": "docker compose down --rmi local --remove-orphans",
     "docker:wipe": "docker compose down -v --remove-orphans",
     "docker:reset": "npm run docker:wipe && npm run docker:up && npm run init",
-    "access-db": "docker exec -it postgres psql -U postgres -d aegis_lol_storage",
+    "db-init": "tsx src/migrate.ts",
+    "db-access": "docker exec -it postgres psql -U postgres -d aegis_lol_storage",
     "test": "mocha --timeout 5000 -r ts-node/register test/**/*.test.ts",
     "lint": "eslint .",
     "prepare": "husky"
